Update GS sync.
