+++ src/layman/__init__.py
@@ -50,12 +50,13 @@ if settings.LAYMAN_REDIS.get(LAYMAN_DEPS_ADJUSTED_KEY) != 'done':
         app.logger.info(f'Adjusting GeoServer')
         with app.app_context():
             # TODO Clear
-            from layman.common.geoserver import ensure_role, ensure_user, ensure_user_role, ensure_wms_srs_list, ensure_proxy_base_url
+            from layman.common.geoserver import sync_all_users, ensure_role, ensure_user, ensure_user_role, ensure_wms_srs_list, ensure_proxy_base_url
             if settings.GEOSERVER_ADMIN_AUTH:
                 ensure_role(settings.LAYMAN_GS_ROLE)
                 ensure_user(settings.LAYMAN_GS_USER, settings.LAYMAN_GS_PASSWORD)
                 ensure_user_role(settings.LAYMAN_GS_USER, 'ADMIN')
                 ensure_user_role(settings.LAYMAN_GS_USER, settings.LAYMAN_GS_ROLE)
+                sync_all_users()
             ensure_wms_srs_list([int(srs.split(':')[1]) for srs in settings.INPUT_SRS_LIST])
             if settings.LAYMAN_GS_PROXY_BASE_URL != '':
                 ensure_proxy_base_url(settings.LAYMAN_GS_PROXY_BASE_URL)
