+++ src/layman/authz/read_everyone_write_everyone.py
@@ -1,6 +1,7 @@
 from flask import g, request
 
 from layman import LaymanError
+from layman.common import geoserver as gs
 
 
 def authorize():
@@ -13,3 +14,11 @@ def get_publication_access_rights(publ_type, username, publication_name):
     return {
         'guest': 'w',
     }
+
+
+def get_GS_roles(username, type):
+    if type == 'r':
+        roles = gs.get_roles_anyone(username)
+    elif type == 'w':
+        roles = gs.get_roles_anyone(username)
+    return roles
