+++ src/layman/authz/read_everyone_write_owner.py
@@ -3,6 +3,8 @@ from flask import g, request
 from layman import LaymanError
 from layman.authn.url_path import USER_PATH_PATTERN
 
+from layman.common import geoserver as gs
+
 
 def authorize():
     if request.method in ['GET']:
@@ -32,3 +34,11 @@ def get_publication_access_rights(publ_type, username, publication_name):
     return {
         'guest': 'r',
     }
+
+
+def get_GS_roles(username, type):
+    if type == 'r':
+        roles = gs.get_roles_anyone(username)
+    elif type == 'w':
+        roles = gs.get_roles_owner(username)
+    return roles
